examples/tensorflow2_keras/requirements-macos.txt
tensorflow-macos
pillow
vtsserving>=1.0.0
